requirements.txt
tqdm
numpy
opencv-python
python-rtmidi
cursor